fix PR comments batch #2
