+++ libs/model/src/lib/utils.ts
@@ -52,6 +52,21 @@ export function titleCase(text: string) {
   return text[0].toUpperCase() + text.substring(1);
 }
 
+/**
+ * Trims a string without cropping the last word if keepLastWordComplete is set to true
+ * Origin : libs\utils\src\lib\pipes\max-length.pipe.ts
+ */
+ export function trimString(string: string, length: number, keepLastWordComplete?: boolean) {
+  if (!string?.length || string.length <= length) return string;
+
+  let trimmedString = string.substr(0, length);
+  if (keepLastWordComplete) {
+    const lastWordIndex = Math.min(trimmedString.length, trimmedString.lastIndexOf(' '));
+    trimmedString = trimmedString.substr(0, lastWordIndex);
+  }
+  return `${trimmedString}...`;
+}
+
 export function displayName(person: Person) {
   return `${titleCase(person.firstName).trim()} ${titleCase(person.lastName).trim()}`;
 }
